Migrate to paging 3.0
