+++ fhgapp/build.gradle
@@ -78,7 +78,7 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
     implementation 'androidx.lifecycle:lifecycle-common-java8:2.3.1'
-    implementation 'androidx.paging:paging-runtime-ktx:2.1.0'
+    implementation 'androidx.paging:paging-runtime-ktx:3.0.0'
 
     implementation "com.jakewharton.timber:timber:4.7.1"
     implementation "org.jsoup:jsoup:1.13.1"
@@ -111,7 +111,7 @@ dependencies {
     testImplementation "org.robolectric:robolectric:4.3"
     testImplementation "org.mockito:mockito-core:3.11.0"
     testImplementation 'androidx.arch.core:core-testing:2.1.0'
-    testImplementation 'androidx.paging:paging-common-ktx:2.1.0'
+    testImplementation 'androidx.paging:paging-common-ktx:3.0.0'
 
     androidTestImplementation "com.google.code.findbugs:jsr305:3.0.2"
     androidTestImplementation 'androidx.test:runner:1.3.0'
